Code can be run multiple times

+++ filefinder.py
@@ -11,7 +11,7 @@ for lang in langs:                          # For all element in the index
 
 for dirpath, dirnames, filenames in os.walk("objects"): # Runs through the minecraft assets
     for file in filenames:                              # For each file
-        if file not in codes:                           # If the name of the files aren't in the language hash list
+        if file not in codes and ".json" not in file:                           # If the name of the files aren't in the language hash list
             os.remove(os.path.join(dirpath, file))      # Deletes them
-        else:                                           # Else
+        elif "json" not in file:                        # Else
             os.rename(os.path.join(dirpath, file), os.path.join(dirpath, file) + ".json")   # Renames them in ".json"
